Changed the login flow from login component to login-page. Extracted The current layout to MainLayoutComponent to switch between layouts easily. (#124)
