+++ sonarQuest-frontend/src/app/app-routing/app-routing.module.ts
@@ -1,9 +1,9 @@
-import {EventPageComponent} from './../pages/event-page/event-page.component';
-import {AdminPageComponent} from './../pages/admin-page/admin-page.component';
+import {EventPageComponent} from '../pages/event-page/event-page.component';
+import {AdminPageComponent} from '../pages/admin-page/admin-page.component';
 import {MarketplacePageComponent} from '../pages/marketplace-page/marketplace-page.component';
-import {QuestPageComponent} from './../pages/quest-page/quest-page.component';
+import {QuestPageComponent} from '../pages/quest-page/quest-page.component';
 import {AdventurePageComponent} from '../pages/adventure-page/adventure-page.component';
-import {StartPageComponent} from './../pages/start-page/start-page.component';
+import {StartPageComponent} from '../pages/start-page/start-page.component';
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {MyAvatarPageComponent} from '../pages/my-avatar-page/my-avatar-page.component';
@@ -11,24 +11,30 @@ import {GamemasterPageComponent} from '../pages/gamemaster-page/gamemaster-page.
 import {AuthenticationGuard} from '../login/authentication.guard';
 import {EmptyPageComponent} from '../pages/empty-page/empty-page.component';
 import {RoutingUrls} from './routing-urls';
+import {LoginPageComponent} from "../pages/login-page/login-page.component";
+import {MainLayoutComponent} from "../layouts/main-layout/main-layout.component";
 
 
 const appRoutes: Routes = [
-  {path: '', redirectTo: '/empty', pathMatch: 'full'},
-  {path: RoutingUrls.empty, component: EmptyPageComponent},
-  {path: RoutingUrls.start, component: StartPageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.myAvatar, component: MyAvatarPageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.adventures, component: AdventurePageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.quests, component: QuestPageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.marketplace, component: MarketplacePageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.gamemaster, component: GamemasterPageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.admin, component: AdminPageComponent, canActivate: [AuthenticationGuard]},
-  {path: RoutingUrls.events, component: EventPageComponent, canActivate: [AuthenticationGuard]}
+  {path: '', component: MainLayoutComponent, canActivateChild: [AuthenticationGuard], children: [
+      {path: '', redirectTo: RoutingUrls.myAvatar, pathMatch: 'full'},
+      {path: RoutingUrls.start, component: StartPageComponent},
+      {path: RoutingUrls.myAvatar, component: MyAvatarPageComponent},
+      {path: RoutingUrls.adventures, component: AdventurePageComponent},
+      {path: RoutingUrls.quests, component: QuestPageComponent},
+      {path: RoutingUrls.marketplace, component: MarketplacePageComponent},
+      {path: RoutingUrls.gamemaster, component: GamemasterPageComponent},
+      {path: RoutingUrls.admin, component: AdminPageComponent},
+      {path: RoutingUrls.events, component: EventPageComponent}
+    ]
+  },
+  {path: RoutingUrls.login, component: LoginPageComponent},
+  {path: RoutingUrls.empty, component: EmptyPageComponent}
 ];
 
 @NgModule({ 
   imports: [
-    RouterModule.forRoot(appRoutes)
+    RouterModule.forRoot(appRoutes, { onSameUrlNavigation: 'reload' })
   ],
   exports: [
     RouterModule
