+++ sonarQuest-frontend/src/app/app.component.html
@@ -1,108 +1,1 @@
-<td-layout-nav-list #navList toolbarTitle="SonarQuest" navigationRoute="/"
-                    [opened]="media.registerQuery('gt-sm') | async"
-                    [mode]="(media.registerQuery('gt-sm') | async) ? 'side' : 'over'"
-                    [sidenavWidth]="(media.registerQuery('gt-xs') | async) ? '300px' : '100%'">
-
-  <div td-sidenav-toolbar-content>
-    <i class="ra ra-sword"></i>
-  </div>
-  <mat-nav-list td-sidenav-content (click)="!media.query('gt-sm') && navList.close()">
-    <mat-list-item disableRipple *ngIf="isWorldSelectVisible">
-      <i mat-list-icon class="ra ra-mountains ra-lg navlist"></i>
-      <div flex></div>
-      <mat-select mat-line [placeholder]="currentWorld?.name" [(ngModel)]="currentWorld" floatPlaceholder="never">
-        <mat-option flex="20" *ngFor="let world of worlds" [value]="world" (click)="updateWorld(world)">
-          {{ world.name }}
-        </mat-option>
-      </mat-select>
-      <div></div>
-      <mat-divider mat-inset></mat-divider>
-    </mat-list-item>
-    <div *ngIf="isMyAvatarVisible">
-      <a mat-list-item routerLink="/{{myAvatarUrl}}">
-        <i mat-list-icon class="ra ra-muscle-up ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.MY_AVATAR"| translate}} </h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    <div *ngIf="isAdventuresVisible">
-      <a mat-list-item routerLink="/{{adventuresUrl}}">
-        <i mat-list-icon class="ra ra-blade-bite ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.ADVENTURES"| translate}} </h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    <div *ngIf="isQuestsVisible">
-      <a mat-list-item routerLink="/{{questsUrl}}">
-        <i mat-list-icon class="ra ra-scroll-unfurled ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.QUESTS"| translate}}</h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    <div *ngIf="isMarketplaceVisible">
-      <a mat-list-item routerLink="/{{marketplaceUrl}}">
-        <i mat-list-icon class="ra ra-capitol ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.MARKETPLACE"| translate}} </h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    <div *ngIf="isGamemasterVisible">
-      <a mat-list-item routerLink="/{{gamemasterUrl}}">
-        <i mat-list-icon class="ra ra-forging ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.GAMEMASTER"| translate}} </h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    <div *ngIf="isAdminVisible">
-      <a mat-list-item routerLink="/{{adminUrl}}">
-        <i mat-list-icon class="ra ra-queen-crown ra-lg navlist"></i>
-        <h3 mat-line>{{"APP_COMPONENT.ADMIN"| translate}} </h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-    
-    <div *ngIf="isEventVisible">
-      <a mat-list-item routerLink="/{{eventUrl}}">
-        <i mat-list-icon class="ra ra-book ra-lg navlist"></i>
-        <h3 mat-line>Events</h3>
-      </a>
-      <mat-divider mat-inset></mat-divider>
-    </div>
-
-  </mat-nav-list>
-  <div td-toolbar-content layout="row" flex>
-    <div class="vertical-align">{{determinePageTitle(router.url)}}</div>
-    <div flex></div>
-    <div class="username-toolbar vertical-align">
-        <button *ngIf="user" mat-icon-button routerLink="/{{myAvatarUrl}}"
-                [matTooltip]="'APP_COMPONENT.USER_PROFILE' | translate"
-                [matTooltipClass]="'tooltipMultiline'"
-                matTooltipPosition="below">
-          <mat-icon>person</mat-icon>
-          <span>{{user?.username}}</span>
-        </button>
-    </div>
-    <div flex="5"></div>
-    <div class="lightning-login-align">
-      <button mat-icon-button (click)="toggleDesign()"
-              [matTooltip]="'APP_COMPONENT.BACKGROUND' | translate"
-              [matTooltipClass]="'tooltipMultiline'"
-              matTooltipPosition="below">
-        <i mat-list-icon class="ra ra-lightning-bolt ra-2x navlist"></i>
-      </button>
-      <button *ngIf="!user" mat-icon-button (click)="login()"
-              [matTooltip]="'APP_COMPONENT.LOGIN' | translate"
-              [matTooltipClass]="'tooltipMultiline'"
-              matTooltipPosition="below">
-        <i mat-list-icon class="ra ra-bridge ra-2x navlist"></i>
-      </button>
-      <button *ngIf="user" mat-icon-button (click)="logout()"
-              [matTooltip]="'APP_COMPONENT.LOGOUT' | translate"
-              [matTooltipClass]="'tooltipMultiline'"
-              matTooltipPosition="below">
-        <mat-icon>power_settings_new</mat-icon>
-      </button>
-    </div>
-  </div>
-  <router-outlet></router-outlet>
-</td-layout-nav-list>
+<router-outlet></router-outlet>
