+++ sonarQuest-frontend/src/app/app.component.spec.ts
@@ -1,13 +1,21 @@
-import { TestBed, async } from '@angular/core/testing';
-
-import { AppComponent } from './app.component';
+import {async, TestBed} from '@angular/core/testing';
+import {RouterTestingModule} from '@angular/router/testing';
+import {AppComponent} from './app.component';
+import {TranslateModule, TranslateService} from "@ngx-translate/core";
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [
+        RouterTestingModule,
+        TranslateModule.forRoot()
+      ],
       declarations: [
         AppComponent
       ],
+      providers: [
+        TranslateService
+      ]
     }).compileComponents();
   }));
