+++ sonarQuest-frontend/src/app/login/authentication.guard.ts
@@ -1,24 +1,25 @@
 import {Injectable} from '@angular/core';
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import {ActivatedRouteSnapshot, CanActivateChild, Router, RouterStateSnapshot} from '@angular/router';
 import {AuthenticationService} from './authentication.service';
 import {PermissionService} from '../services/permission.service';
+import {RoutingUrls} from "../app-routing/routing-urls";
 
 @Injectable()
-export class AuthenticationGuard implements CanActivate {
+export class AuthenticationGuard implements CanActivateChild {
 
   constructor(private router: Router,
               private authService: AuthenticationService,
               private permissionService: PermissionService) {
   }
 
-  canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Promise<boolean> {
+  canActivateChild(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Promise<boolean> {
         return this.authenticate(next.url.toString());
   }
 
   authenticate(nextUrl: string): boolean | Promise<boolean> {
     const isLoggedIn = this.authService.isLoggedIn();
     if (!isLoggedIn) {
-      this.router.navigate(['/'], {skipLocationChange: true});
+      this.router.navigate([RoutingUrls.login], { skipLocationChange: true});
       return false;
     }
