+++ sonarQuest-frontend/src/app/pages/login-page/login-page.component.ts
@@ -1,0 +1,29 @@
+import {Component, OnInit} from '@angular/core';
+import {AuthenticationService} from "../../login/authentication.service";
+
+@Component({
+  selector: 'sq-login-page',
+  templateUrl: './login-page.component.html',
+  styleUrls: ['./login-page.component.scss']
+})
+export class LoginPageComponent implements OnInit {
+
+  username = '';
+  password = '';
+
+  constructor(private authService: AuthenticationService) {
+  }
+
+  ngOnInit() {
+  }
+
+  onSubmit() {
+    this.authService.login(this.username, this.password);
+  }
+
+  onKeyDown(event) {
+    if (event.key === "Enter" && this.username != '' && this.password != '') {
+      this.onSubmit();
+    }
+  }
+}
