+++ sonarQuest-frontend/src/index.html
@@ -14,10 +14,10 @@
     var global = window;
   </script>
 </head>
-<body class="background-image">
+<body>
   <app-root>
       <div class="loading">
-          <img src="assets/images/splash_screen.svg" style="height:30%">
+          <img src="assets/images/splash_screen.svg" style="height:30%" alt="This is the start screen of the app.">
           <h1>prepare for...SonarQuest</h1>
           <div class="spinner">
               <div class="double-bounce1"></div>
