+++ sonarQuest-frontend/src/styles.scss
@@ -64,3 +64,14 @@
 }
 
 
+.login-form .mat-focused .mat-form-field-label,
+.login-form .mat-form-field-label,
+.login-form .mat-form-field-label .mat-form-field-required-marker {
+  color: white !important;
+}
+
+.login-form .mat-form-field-ripple {
+  background-color: white !important;
+}
+
+
